Settings: reorganization
Reorganize settings categories to be more user friendly. In addition,
there is now an "Advanced mode". This mode is on by default for
userdebug and eng builds, and can be overlayed to be turned on by
default for user builds. Currently this toggle lives in About phone for
easy discoverability.

Signed-off-by: Roman Birg <roman@cyngn.com>

Conflicts:

	AndroidManifest.xml
	res/values/cm_strings.xml
	res/values/config.xml
	res/xml/button_settings.xml
	res/xml/dashboard_categories.xml
	res/xml/device_info_settings.xml
	res/xml/display_settings.xml
	res/xml/performance_settings.xml
	res/xml/privacy_settings_cyanogenmod.xml
	src/com/android/settings/ButtonSettings.java
	src/com/android/settings/DisplaySettings.java
	src/com/android/settings/SettingsActivity.java
	src/com/android/settings/cyanogenmod/PerformanceSettings.java
	src/com/android/settings/search/IndexDatabaseHelper.java

Change-Id: I18fa8c3e3ee7efc9fcaa7384d5b1cfb2495843f1